refactor(components): move login button outside of header folder
