+++ packages/nextjs/components/index.ts
@@ -1,3 +1,4 @@
 export { Header } from "./Header";
 export { Footer } from "./Footer";
 export { LargeLogo } from "./LargeLogo";
+export { LoginButton } from "./LoginButton";
