+++ packages/nextjs/pages/profile/index.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { useAccount } from "wagmi";
-import { LoginButton } from "~~/components/Header/LoginButton";
+import { LoginButton } from "~~/components";
 import { Order, Plan, TokenAddress, tokenAddressMap } from "~~/context/Types";
 import { useBBContractReads } from "~~/hooks/Botblock";
 import { ContractNames } from "~~/hooks/Botblock/hooksUtils";
